requirements.txt
spheral
pandas
numpy
scipy
matplotlib
plotly
scikit-learn
torch
torchvision
tensorboard
umap-learn[plot]
tqdm
dash
python-dotenv